refactor/Apagando arquivos que não estão mais sendo usados e refatorando Página de compra finalizada
