+++ src/components/ScreenWithHeader.js
@@ -1,11 +1,13 @@
 import styled from "styled-components";
 import tw from "twin.macro";
-import Header from "./Header";
+import Header from "./Header/Header";
+import SubHeader from "./SubHeader/SubHeader";
 
 export default function ScreenWithHeader({ children }) {
   return (
     <Container>
       <Header />
+      <SubHeader />
       {children}
     </Container>
   );
