+++ src/components/SearchBar/styled.js
