+++ src/pages/CompletedPurchase/styled.js
@@ -1,0 +1,25 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+font-family: 'Roboto', sans-serif;
+display: flex;
+width: 100%;
+padding-top: 6rem;
+`;
+
+export const PurchaseMessageContainer = styled.div`
+display: flex;
+flex-direction: column;
+justify-content: center;
+align-items: center;
+border: solid;
+border-color: rgb(229 231 235);
+width: 65%;
+`;
+
+export const UserData = styled.div`
+display: flex;
+flex-direction: column;
+padding-top: 10px;
+gap: 5px;
+`;
