+++ src/pages/Listagem/Listagem.js
@@ -1,6 +1,6 @@
 import { Wrapper } from "./styled";
-import Header from "../../component/Header/Header";
-import SubHeader from "../../component/SubHeader/SubHeader";
+import Header from "../../components/Header/Header";
+import SubHeader from "../../components/SubHeader/SubHeader";
 
 export default function Listagem() {
   return (
